feat: Moves types into a type file and creates a new fetch call for getting animals by location
