+++ src/Components/AnimalDetails/AnimalDetails.tsx
@@ -1,40 +1,5 @@
 import React from "react";
-
-type Animal = {
-  id: number;
-  primary_photo_cropped: {
-    full: string,
-    large: string,
-    medium: string,
-    small: string
- } | null;
- url: string,
-  age: string,
-  breeds: {
-    mixed: boolean,
-    primary: string,
-    secondary: string,
-    unknown: boolean
-  },
-  contact: {
-    address: {
-      address1: string,
-      address2: string,
-      city: string,
-      country: string,
-      postcode: string,
-      state: string
-    }
-    email: string,
-    phone: string
-  }
-  description: string,
-  gender: string,
-  name: string,
-  size: string,
-  species: string,
-  type: string,
-};
+import { Animal } from '../../types';
 
 interface AnimalProps {
   animals: Animal[]
