+++ src/types.d.ts
@@ -1,0 +1,37 @@
+type Animal = {
+  id: number;
+  primary_photo_cropped: {
+    full: string,
+    large: string,
+    medium: string,
+    small: string
+ } | null;
+ url: string,
+  age: string,
+  breeds: {
+    mixed: boolean,
+    primary: string,
+    secondary: string,
+    unknown: boolean
+  },
+  contact: {
+    address: {
+      address1: string,
+      address2: string,
+      city: string,
+      country: string,
+      postcode: string,
+      state: string
+    }
+    email: string,
+    phone: string
+  }
+  description: string,
+  gender: string,
+  name: string,
+  size: string,
+  species: string,
+  type: string,
+};
+
+export { Animal };
